requirements.txt
# CLI
click>=8.1

# Web-Crawling
requests>=2.31
beautifulsoup4>=4.12

# PDF
PyMuPDF>=1.23

# Word (.docx)
python-docx>=1.1

# Excel (.xlsx)
openpyxl>=3.1

# PowerPoint (.pptx)
python-pptx>=0.6

# Bilder (OCR)
Pillow>=10.0
pytesseract>=0.3

# KI-Analyse
anthropic>=0.40

# Konfiguration
python-dotenv>=1.0
